requirements.txt
mss
Pillow
transformers
torch
nudenet
onnxruntime
huggingface_hub
